Split modules into different entry points, separate vendor modules
