+++ views/layout.html
@@ -23,6 +23,8 @@
 
 <!-- javascripts -->
 <%= javascript(
+  'src/vendor.js',
+  'src/mailpoet.js',
   'src/admin.js'
 )%>
